Embedded 0.0.7 - Removed static offset for AD7124, replaced with barometric sensor reading - AD7124 sample rate adjusted - Data logging and export functionality through RAM buffer - Fixed return values in pressure.h/.c - Added sample rate hardware timer on htim2 - Added software filter on AD7124, adjusted hardware filter setting - Added reset wait time to barometric pressure init - Scaffolding in place for flash based data storage, not yet used - Forced chipid for AD7124 to 0 to prevent confusion - Fixed missing return values in POST, not yet implemented - Created interrupt versions of OctalSpiFlashErase commands, not yet used - Remove unused code. misc code cleanup
git-svn-id: file:///Users/whitneyjackson/backup_repo@28 8778e90b-af49-4903-8128-3a2f383b7c61

+++ trunk/CUR_AdaptiveTherapyMonitor/Core/Inc/pressure.h
@@ -33,7 +33,7 @@ typedef enum
 	ePressureUnits_psi
 } ePressureUnits;
 
-uint32_t GetBaselinePressure(void);
-uint32_t GetPressure(ePressureUnits units);
+float GetBaselinePressure(void);
+float GetPressure(ePressureUnits units);
 
 #endif /* INC_PRESSURE_H_ */
